server/requirements.txt
flask
flask-admin
flask-simplelogin
flask-bootstrap
flask-sqlalchemy
sqlalchemy-serializer
dynaconf
click
flask-restful
flask-cors
